Extract CurrentPicker date component
