+++ frontend/src/metabase/query_builder/components/filters/pickers/DatePicker/CurrentPicker/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CurrentPicker";
