+++ frontend/src/metabase/query_builder/components/filters/pickers/DatePicker/RelativeDatePicker.styled.tsx
@@ -22,46 +22,15 @@ export const NumericInput = styled(BaseNumericInput)<BaseProps>`
   }
 `;
 
-type ButtonProps = {
-  primaryColor?: string;
-  selected?: boolean;
-};
-
-export const CurrentButton = styled(Button)<ButtonProps>`
-  border: none;
-  border-radius: 99px;
-
-  background-color: ${({ selected, primaryColor = color("brand") }) =>
-    selected ? primaryColor : alpha(primaryColor, 0.1)};
-  color: ${({ selected, primaryColor = color("brand") }) =>
-    selected ? "white" : primaryColor};
-
-  padding-top: ${space(1)};
-  padding-bottom: ${space(1)};
-
-  &:hover {
-    color: white;
-    background-color: ${props => props.primaryColor || color("brand")};
-  }
-`;
-
-export const CurrentContainer = styled.div<{ first?: boolean }>`
-  display: flex;
-  flex-wrap: no-wrap;
-  grid-gap: ${space(2)};
-  margin-bottom: ${({ first }) => (first ? space(2) : "")};
-`;
-
-export const CurrentPopover = styled.div`
-  color: ${color("white")};
-  background-color: ${color("black")};
-  padding: ${space(1)} ${space(2)};
-`;
-
 export const OptionsContainer = styled.div`
   background-color: ${color("white")};
   padding: ${space(2)} ${space(1)};
 `;
+
+type ButtonProps = {
+  primaryColor?: string;
+  selected?: boolean;
+};
 
 type OptionButtonProps = ButtonProps & {
   reverseIconDirection?: boolean;
