+++ frontend/src/metabase/query_builder/components/filters/pickers/DatePicker/RelativeDatePicker.tsx
@@ -5,7 +5,6 @@ import moment from "moment";
 import { assoc } from "icepick";
 
 import {
-  formatBucketing,
   formatStartingFrom,
   getRelativeDatetimeInterval,
   getStartingFrom,
@@ -18,9 +17,6 @@ import TippyPopover from "metabase/components/Popover/TippyPopover";
 
 import Filter from "metabase-lib/lib/queries/structured/Filter";
 import {
-  CurrentButton,
-  CurrentContainer,
-  CurrentPopover,
   GridContainer,
   GridText,
   MoreButton,
@@ -42,75 +38,15 @@ export const NextPicker = (props: Props) => (
   />
 );
 
-const periodPopoverText = (period: string) => {
-  const now = moment();
-  let start: string, end: string;
-  switch (period) {
-    case "day":
-      return t`Right now, this is ${now.format("ddd, MMM D")}`;
-    case "week":
-      start = now.startOf("week").format("ddd, MMM D");
-      end = now.endOf("week").format("ddd, MMM D");
-      return t`Right now, this is ${start} - ${end}`;
-    case "month":
-      start = now.startOf("month").format("ddd, MMM D");
-      end = now.endOf("month").format("ddd, MMM D");
-      return t`Right now, this is ${start} - ${end}`;
-    case "quarter":
-      start = now.startOf("quarter").format("ddd, MMM D");
-      end = now.endOf("quarter").format("ddd, MMM D");
-      return t`Right now, this is ${start} - ${end}`;
-    case "year":
-      start = now.startOf("year").format("MMM D, YYYY");
-      end = now.endOf("year").format("MMM D, YYYY");
-      return t`Right now, this is ${start} - ${end}`;
-  }
+const CURRENT_INTERVAL_NAME = {
+  day: t`today`,
+  week: t`this week`,
+  month: t`this month`,
+  year: t`this year`,
+  quarter: t`this quarter`,
+  minute: t`this minute`,
+  hour: t`this hour`,
 };
-
-type CurrentPickerProps = {
-  className?: string;
-  filter: Filter;
-  primaryColor?: string;
-  onCommit: (filter?: any[]) => void;
-};
-
-export function CurrentPicker(props: CurrentPickerProps) {
-  const {
-    className,
-    primaryColor,
-    onCommit,
-    filter: [operator, field, _intervals, unit],
-  } = props;
-  return (
-    <div className={className}>
-      {DATE_PERIODS.map((periods, index) => (
-        <CurrentContainer key={periods.length} first={index === 0}>
-          {periods.map(period => (
-            <TippyPopover
-              key={period}
-              placement="bottom"
-              delay={[500, null]}
-              content={
-                <CurrentPopover>{periodPopoverText(period)}</CurrentPopover>
-              }
-            >
-              <CurrentButton
-                key={period}
-                primaryColor={primaryColor}
-                selected={operator && unit === period.toLowerCase()}
-                onClick={() => {
-                  onCommit([operator, field, "current", period]);
-                }}
-              >
-                {formatBucketing(period, 1)}
-              </CurrentButton>
-            </TippyPopover>
-          ))}
-        </CurrentContainer>
-      ))}
-    </div>
-  );
-}
 
 export const DATE_PERIODS = [
   ["day", "week", "month"],
@@ -143,16 +79,6 @@ const getStartingFromUnits = (
 
 const getCurrentString = (filter: Filter) =>
   t`Include ${getCurrentIntervalName(filter)}`;
-
-const CURRENT_INTERVAL_NAME = {
-  day: t`today`,
-  week: t`this week`,
-  month: t`this month`,
-  year: t`this year`,
-  quarter: t`this quarter`,
-  minute: t`this minute`,
-  hour: t`this hour`,
-};
 
 function getCurrentIntervalName(filter: Filter) {
   if (filter[0] === "time-interval") {
